move MesosTaskConfig to its own module
